+++ src/session/user_list.rs
@@ -98,6 +98,6 @@ impl UserList {
         let priv_ = imp::UserList::from_instance(self);
         let list = priv_.list.borrow();
         let position = list.len() - 1;
-        self.items_changed(position as u32, 0, 1 as u32);
+        self.items_changed(position as u32, 0, 1);
     }
 }
